test: merge moon/planet collision regression tests into one file
Both were the same shape (date-based conjunction regression check
against renderSolarSystem positions) and only existed as separate
files because they were added one at a time.
